+++ tools/amavlink/amavlink/AMavlinkErrors.py
@@ -41,8 +41,10 @@ class AMavlinkHeartbeatNotReceivedError(AMavLinkError):
 class AMavlinkCLIParseError(AMavLinkError):
     pass
 
+
 class AMavlinkCLIParamVerificationError(AMavLinkError):
     pass
+
 
 class AMavlinkParamFileDelimiterNotFoundInLineError(AMavLinkError):
     pass